% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_iso.r
\name{in_get_category_difference}
\alias{in_get_category_difference}
\title{Get countries that are in category1 but not in category2}
\usage{
in_get_category_difference(category1, category2)
}
\arguments{
\item{category1}{Character, first category name}

\item{category2}{Character, second category name}
}
\value{
Character vector of ISO3 codes present in category1 but not in category2
}
\description{
Get countries that are in category1 but not in category2
}
\examples{
eu_not_nato <- get_category_difference("EU", "NATO")

}
\keyword{internal}
