% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_print_debug.r
\name{in_print_debug}
\alias{in_print_debug}
\title{Helper function for nice debug/verbose printing}
\usage{
in_print_debug(
  text,
  verbose = TRUE,
  debug = FALSE,
  type = "info",
  text_type = "INFO"
)
}
\description{
Helper function for nice debug/verbose printing
}
\keyword{internal}
