% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{METADATA_Q}
\alias{METADATA_Q}
\title{Metadata for Quarterly Variables}
\format{
A data frame with the following columns:
\describe{
\item{Origin}{Short identifier for the data source}
\item{Reference}{Detailed source of the data}
\item{Symbol}{Variable identifier}
\item{Indicator}{Description of the variable}
\item{Frequency}{Always "Q" for this dataset}
}
}
\usage{
data(METADATA_Q)
}
\description{
A dataset containing information about all available quarterly variables.
}
\keyword{datasets}
