% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{METADATA_Y}
\alias{METADATA_Y}
\title{Metadata for Yearly Variables}
\format{
A data frame with the following columns:
\describe{
\item{Origin}{Source of the data}
\item{Symbol}{Variable identifier}
\item{Indicator}{Description of the variable}
\item{Frequency}{Always "Y" for this dataset}
}
}
\usage{
data(METADATA_Y)
}
\description{
A dataset containing information about all available yearly variables.
}
\keyword{datasets}
