% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_info.r
\name{wp_info}
\alias{wp_info}
\title{Display information about available data}
\usage{
wp_info(frequency = NULL)
}
\arguments{
\item{frequency}{Character, either "Q", "Y", or NULL for both}
}
\value{
Invisibly returns the metadata for the specified frequency/frequencies
}
\description{
Display information about available data
}
