% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPD-package.r
\docType{package}
\name{WPD}
\alias{WPD}
\alias{WPD-package}
\title{WPD: World Panel Data Visualization and Analysis Toolkit}
\description{
A comprehensive package for accessing, analyzing and visualizing global economic
and social indicators. The package provides tools for working with panel data,
with built-in visualization capabilities and data analysis functions.
}
\section{Main Functions}{

\itemize{
\item \code{\link{wp_data}}: Load and filter panel data
\item \code{\link{wp_plot_series}}: Create time series plots
\item \code{\link{wp_plot_bar}}: Create bar charts
\item \code{\link{wp_plot_scatter}}: Create scatter plots
\item \code{\link{wp_info}}: Get information about available data
}
}

\section{Country Code Utilities}{

\itemize{
\item \code{\link{wp_ctry2iso}}: Convert country names to ISO3 codes
\item \code{\link{wp_from_iso}}: Convert ISO codes to country information
\item \code{\link{wp_get_category}}: Get countries belonging to specific categories
}
}

\section{Data}{

The package includes:
\itemize{
\item Annual and quarterly macroeconomic indicators
\item Cross-country panel data
\item Built-in seasonal adjustment capabilities
\item Efficient on-demand data loading
}
}

\references{
Package website: \url{https://benjaminpeeters.github.io/WPD}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://benjaminpeeters.github.io/WPD/}
  \item Report bugs at \url{https://github.com/benjaminpeeters/WPD/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Peeters \email{benjaminpeeters@protonmail.com} (\href{https://orcid.org/0000-0002-5095-2021}{ORCID})

}
\keyword{internal}
