% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_iso.r
\name{in_get_ctry_categories}
\alias{in_get_ctry_categories}
\title{Get all categories that contain specific countries}
\usage{
in_get_ctry_categories(country_codes)
}
\arguments{
\item{country_codes}{Character vector of ISO3 country codes}
}
\value{
Character vector of category names
}
\description{
Get all categories that contain specific countries
}
\keyword{internal}
